RamseyGrowthModel.jl

Defining and Solving a Growth Model

RamseyGrowthModel.GrowthModel — Type

Defines a Ramsey Growth Model.

Fields

β - discount factor

δ - depreciation rate on capital

u - utility function

f - production function

Constructors

If you want to use a CRRA utility function and a Cobb-Douglas production function, use the constructor:

GrowthModel(β::Float64, δ::Float64, γ::Float64, α::Float64, A::Float64)

where γ is a parameter for sample_u, and α and A are parameters for sample_f.

RamseyGrowthModel.solve — Function

Returns the best possible capital and consumption allocation (as a DataFrame).

solve(model::GrowthModel, T::Int64, K₀::Float64; kwargs...)

Arguments

model - previously defined growth model

T - considered time horizon

K₀ - initial capital

Keyword Arguments

The algorithm uses a binary search; if you want, you can override the default maximum number of iterations (max_iter=1000) or error tolerance (tol=K₀/1e6).

Sample Functions Used in the Model

RamseyGrowthModel.sample_u — Function

Returns CRRA (constant relative risk aversion) utility function $u(c)$ with a given $γ$.

\[u(c) = \frac{c^{1 - γ} - 1}{1 - γ} + 1\]

What does it look like? See for yourself here.

RamseyGrowthModel.sample_f — Function

Returns per-capita Cobb-Douglas production function $f(k)$ with given $A$ and $α$.

\[f(k) = A \cdot k^α\]